+++ src/main/java/ssrahoo/marketplaceapi/dto/TransactionResponseDto.java
@@ -1,5 +1,6 @@
 package ssrahoo.marketplaceapi.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -9,7 +10,7 @@ public record TransactionResponseDto(
         String productName,
         UUID productId,
         Integer amount,
-        Double totalPrice,
+        BigDecimal totalPrice,
         Instant created,
         Instant modified
 ) {}
